requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.2
djoser>=2.3
psycopg>=3.1
python-dotenv>=1.0
dj-database-url
